refactor(backend): split market_calendar into a package by audience
The module had reached 438 lines, ~40% of it venue trait data. Now:

- listings.py — pure data (SUFFIX_LISTINGS, INDEX_CALENDARS,
  EXTENDED_HOURS, crypto quote sets): adding a venue is a data edit that
  cannot change behavior
- venue.py — the Venue facade + per-calendar cache
- symbol.py — the Symbol str subclass (the one ticker-shape parser)
- schedule.py — scheduler-facing helpers (any_venue_open,
  schedule_poll_hint) with their deliberately opposite failure postures
  documented side by side
- __init__.py re-exports the public API: every existing import
  (`from app.services.market_calendar import Symbol`) works verbatim

Dependencies flow one way (schedule → symbol → venue → listings). No
behavior change; 773 tests pass untouched.
