+++ CLAUDE.md
@@ -83,7 +83,7 @@ Two long-lived branches: `main` (production, `fibenchi:latest`) and `dev` (stagi
   - `pseudo_etf.py` — Equal-weight performance with quarterly rebalancing. Two modes: `_calc_static` (all constituents from day 1) and `_calc_dynamic` (assets join when price ≥ threshold, prevents penny-stock distortion).
   - `portfolio.py` — Portfolio index using dynamic entry with `min_entry_price=10.0`.
   - `utils.py` — Shared `prices_to_df()` converting ORM objects to indexed DataFrame.
-- `services/market_calendar.py` — `Symbol` (a `str` subclass, drop-in for plain ticker strings) resolves its own venue: `Symbol("IWDA.AS").venue` → a cached per-calendar `Venue` facade over `exchange_calendars` (suffix + index tables; unmapped → `None`); ticker shape is only ever interpreted in `Symbol.calendar_name`. `Venue` answers session questions for the σ-Move gap guard (`session_dates_for_index` → `compute_indicators(df, session_dates=...)`, so holidays aren't mistaken for feed holes) and schedule questions: `is_open`/`next_open`/`next_close`/`previous_close` and `phase()` (`premarket|open|aftermarket|closed`; extended windows are `EXTENDED_HOURS` data anchored to actual open/close, half-day-aware, US-only). Fail-safe: unknown venue or out-of-range query returns `None` and callers fall back to business-day heuristics; the quote feed's `market_state` remains the live authority.
+- `services/market_calendar/` — package split by audience: `listings.py` (pure venue trait data: `SUFFIX_LISTINGS` with calendar+currency per suffix, index table, `EXTENDED_HOURS`), `symbol.py`, `venue.py`, `schedule.py` (`any_venue_open` gate, `schedule_poll_hint`); public API re-exported from `__init__`. `Symbol` (a `str` subclass, drop-in for plain ticker strings) resolves its own venue: `Symbol("IWDA.AS").venue` → a cached per-calendar `Venue` facade over `exchange_calendars` (unmapped → `None`); `Symbol.currency` infers venue currency; ticker shape is only ever interpreted in `Symbol`. `Venue` answers session questions for the σ-Move gap guard (`session_dates_for_index` → `compute_indicators(df, session_dates=...)`, so holidays aren't mistaken for feed holes) and schedule questions: `is_open`/`next_open`/`next_close`/`previous_close` and `phase()` (`premarket|open|aftermarket|closed`; extended windows are `EXTENDED_HOURS` data anchored to actual open/close, half-day-aware, US-only). Fail-safe: unknown venue or out-of-range query returns `None` and callers fall back to business-day heuristics; the quote feed's `market_state` remains the live authority.
 - `services/entity_lookups.py` — Canonical lookup helpers: `find_asset()` returns `None` (for ephemeral price fetches of untracked symbols like ETF holdings), `get_asset()`/`get_group()`/`get_pseudo_etf()` raise HTTP 404.
 - `utils.py` — `async_threadable` decorator (wraps sync functions for `asyncio.to_thread`), `TTLCache` class.
 - `constants.py` — `PERIOD_DAYS` (period string → calendar days), `WARMUP_DAYS` (derived from max indicator warmup periods).
