--- backend/app/services/market_calendar.py
@@ -1,438 +1,0 @@
-"""Symbol → venue trading-calendar resolution and session/schedule queries.
-
-The entry point is :class:`Symbol` — a ``str`` subclass, so it drops in
-anywhere a plain ticker string is used — whose ``.venue`` resolves the
-symbol's trading venue: ``Symbol("IWDA.AS").venue`` → the XAMS
-:class:`Venue`, shared by every symbol on that calendar. The venue answers
-session and schedule questions: which dates are trading sessions, when the
-venue opens/closes, and which phase (premarket / open / aftermarket /
-closed) an instant falls in.
-
-The primary consumer is the σ-Move gap guard (issue #559) — with real
-session dates (``exchange_calendars``) a missing bar can be distinguished
-from an exchange holiday. The schedule/phase side exists for market-state
-uses (pre/post-market markers, SSE polling cadence).
-
-Structure: ticker shape is only ever interpreted in
-``Symbol.calendar_name``; everything venue-specific beyond that is *data*
-(the suffix/index tables, the ``EXTENDED_HOURS`` table), and all schedule
-questions are answered by the :class:`Venue` object — callers never branch
-on ticker shape themselves.
-
-Everything here is fail-safe by design: an unmapped symbol, an unknown
-calendar, or a query outside the calendar's bounds returns ``None`` and the
-caller falls back to its calendar-less heuristic. A wrong answer from this
-module must never be worse than not having it.
-"""
-
-from __future__ import annotations
-
-import logging
-from dataclasses import dataclass
-from datetime import date, datetime, timedelta, timezone
-from functools import cached_property
-from typing import Any
-
-import pandas as pd
-
-logger = logging.getLogger(__name__)
-
-@dataclass(frozen=True)
-class Listing:
-    """Venue traits of a Yahoo exchange suffix: trading calendar + currency.
-
-    ``calendar`` is None for venues exchange_calendars doesn't model (their
-    symbols keep the business-day fallback); ``currency`` is the *venue*
-    currency, used only as a fallback when Yahoo doesn't report one (see
-    ``resolve_currency`` — subunits like GBp are Yahoo-side and handled there).
-    """
-
-    calendar: str | None
-    currency: str | None
-
-
-# The one venue table: Yahoo suffix (the part after the last '.') →
-# (exchange_calendars name, ISO 4217 currency). Formerly two independently
-# maintained tables (SUFFIX_CALENDARS here, EXCHANGE_CURRENCY_MAP in
-# yahoo/currency.py) over the same key space.
-SUFFIX_LISTINGS: dict[str, Listing] = {
-    # Euronext
-    "AS": Listing("XAMS", "EUR"),   # Amsterdam
-    "BR": Listing("XBRU", "EUR"),   # Brussels
-    "PA": Listing("XPAR", "EUR"),   # Paris
-    "LS": Listing("XLIS", "EUR"),   # Lisbon
-    "IR": Listing("XDUB", "EUR"),   # Dublin
-    # Rest of Europe
-    "DE": Listing("XETR", "EUR"),   # Xetra
-    "F": Listing("XFRA", "EUR"),    # Frankfurt floor
-    "MI": Listing("XMIL", "EUR"),   # Borsa Italiana
-    "L": Listing("XLON", "GBP"),    # London
-    "IL": Listing("XLON", "GBP"),   # London IOB
-    "SW": Listing("XSWX", "CHF"),   # SIX Swiss
-    "MC": Listing("XMAD", "EUR"),   # Madrid
-    "VI": Listing("XWBO", "EUR"),   # Vienna
-    "ST": Listing("XSTO", "SEK"),   # Stockholm
-    "HE": Listing("XHEL", "EUR"),   # Helsinki
-    "OL": Listing("XOSL", "NOK"),   # Oslo
-    "CO": Listing("XCSE", "DKK"),   # Copenhagen
-    "IC": Listing("XICE", "ISK"),   # Iceland
-    "WA": Listing("XWAR", "PLN"),   # Warsaw
-    "PR": Listing("XPRA", "CZK"),   # Prague
-    "BD": Listing("XBUD", "HUF"),   # Budapest
-    "AT": Listing("ASEX", "EUR"),   # Athens
-    "IS": Listing("XIST", "TRY"),   # Istanbul
-    # Middle East & Africa
-    "TA": Listing("XTAE", "ILS"),   # Tel Aviv
-    "SR": Listing("XSAU", "SAR"),   # Saudi (Tadawul)
-    "QA": Listing(None, "QAR"),     # Qatar — no exchange_calendars calendar
-    "JO": Listing("XJSE", "ZAR"),   # Johannesburg
-    # Americas
-    "TO": Listing("XTSE", "CAD"),   # Toronto
-    "V": Listing("XTSE", "CAD"),    # TSX Venture — shares Toronto's schedule
-    "MX": Listing("XMEX", "MXN"),   # Mexico
-    "SA": Listing("BVMF", "BRL"),   # São Paulo
-    "BA": Listing("XBUE", "ARS"),   # Buenos Aires
-    "SN": Listing("XSGO", "CLP"),   # Santiago
-    # Asia-Pacific
-    "T": Listing("XTKS", "JPY"),    # Tokyo
-    "HK": Listing("XHKG", "HKD"),   # Hong Kong
-    "SS": Listing("XSHG", "CNY"),   # Shanghai
-    "SZ": Listing("XSHG", "CNY"),   # Shenzhen — no own calendar; same holidays
-    "KS": Listing("XKRX", "KRW"),   # Korea (KOSPI)
-    "KQ": Listing("XKRX", "KRW"),   # KOSDAQ (same holidays)
-    "TW": Listing("XTAI", "TWD"),   # Taiwan (TWSE)
-    "TWO": Listing("XTAI", "TWD"),  # Taiwan OTC (same holidays)
-    "SI": Listing("XSES", "SGD"),   # Singapore
-    "AX": Listing("XASX", "AUD"),   # Australia
-    "NZ": Listing("XNZE", "NZD"),   # New Zealand
-    "JK": Listing("XIDX", "IDR"),   # Jakarta
-    "BK": Listing("XBKK", "THB"),   # Bangkok
-    # South Asia
-    "NS": Listing("XBOM", "INR"),   # NSE India — XBOM (BSE) shares holidays
-    "BO": Listing("XBOM", "INR"),   # BSE India
-}
-
-# Common Yahoo index symbols → calendar of the venue they track.
-INDEX_CALENDARS: dict[str, str] = {
-    "^AEX": "XAMS",
-    "^BFX": "XBRU",
-    "^FCHI": "XPAR",
-    "^GDAXI": "XETR",
-    "^STOXX50E": "XETR",
-    "^FTSE": "XLON",
-    "^SSMI": "XSWX",
-    "^IBEX": "XMAD",
-    "^GSPC": "XNYS",
-    "^DJI": "XNYS",
-    "^IXIC": "XNYS",
-    "^NDX": "XNYS",
-    "^RUT": "XNYS",
-    "^VIX": "XNYS",
-    "^N225": "XTKS",
-    "^HSI": "XHKG",
-}
-
-# Unsuffixed Yahoo symbols are US listings (NYSE/NASDAQ share the holiday set).
-DEFAULT_US_CALENDAR = "XNYS"
-
-# Quote currencies Yahoo uses in crypto pair symbols (BTC-USD, ETH-EUR, SOL-BTC).
-# A hyphen alone is NOT crypto — US class shares (BRK-B, BF-B) and preferreds
-# (BAC-PL) are hyphenated too; those must fall through to the US default.
-# European share classes carry a suffix after the hyphenated part (NOVO-B.CO),
-# so the '.' branch resolves them before the hyphen could matter.
-CRYPTO_QUOTE_CURRENCIES = frozenset({
-    "USD", "EUR", "GBP", "JPY", "CAD", "AUD", "CHF", "BTC", "ETH", "USDT", "USDC",
-})
-# The fiat subset — a BTC-quoted pair has no meaningful display currency.
-_FIAT_QUOTES = frozenset({"USD", "EUR", "GBP", "JPY", "CAD", "AUD", "CHF"})
-
-
-@dataclass(frozen=True)
-class ExtendedHours:
-    """A venue's extended-trading window, anchored to the regular session.
-
-    ``pre_offset`` — how long before the regular open premarket starts;
-    ``post_offset`` — how long after the regular close aftermarket ends.
-    Anchoring to the *actual* schedule (not wall-clock constants) keeps
-    half-days correct: on a US early-close day (13:00 ET) aftermarket ends
-    17:00 ET, and DST is handled because the calendar's open/close move.
-    """
-
-    pre_offset: timedelta
-    post_offset: timedelta
-
-
-# Venues with a real extended-hours session, keyed by calendar name — a data
-# table, not per-ticker conditionals. US listings: premarket 4:00 ET
-# (= open − 5:30), aftermarket until 20:00 ET (= close + 4:00), matching the
-# window Yahoo's PRE/POST market states cover. European venues have auction
-# phases, not retail extended sessions, so they are deliberately absent.
-EXTENDED_HOURS: dict[str, ExtendedHours] = {
-    "XNYS": ExtendedHours(
-        pre_offset=timedelta(hours=5, minutes=30), post_offset=timedelta(hours=4)
-    ),
-}
-
-
-def _as_utc(at: datetime | None) -> pd.Timestamp:
-    ts = pd.Timestamp(at) if at is not None else pd.Timestamp.now(tz=timezone.utc)
-    return ts.tz_localize(timezone.utc) if ts.tzinfo is None else ts
-
-
-class Venue:
-    """Schedule facade for one trading venue.
-
-    Bundles the exchange_calendars calendar with the venue's traits (extended
-    hours) so callers ask the venue questions instead of re-deriving anything
-    from the symbol. One instance per calendar, shared by all symbols on that
-    venue (see :attr:`Symbol.venue`). All methods return ``None`` on
-    out-of-range or otherwise unanswerable queries — never raise.
-    """
-
-    def __init__(self, name: str, calendar: Any, extended_hours: ExtendedHours | None):
-        self.name = name
-        self.extended_hours = extended_hours
-        self._cal = calendar
-
-    def __repr__(self) -> str:
-        return f"Venue({self.name!r})"
-
-    # -- sessions -----------------------------------------------------------
-
-    def session_dates(self, start: date, end: date) -> set[date] | None:
-        """All trading sessions in [start, end] (clamped to calendar bounds).
-
-        None when the range can't be answered; an empty set is a real answer.
-        """
-        try:
-            first = max(pd.Timestamp(start), self._cal.first_session)
-            last = min(pd.Timestamp(end), self._cal.last_session)
-            if first > last:
-                return None
-            return {ts.date() for ts in self._cal.sessions_in_range(first, last)}
-        except Exception:
-            logger.warning(
-                "Session query failed for %s (%s..%s)", self.name, start, end, exc_info=True
-            )
-            return None
-
-    def session_dates_for_index(self, index) -> set[date] | None:
-        """Sessions covering a price DataFrame's date index (None if unusable).
-
-        Convenience for the indicator pipeline: accepts an index of dates,
-        datetimes, or Timestamps and clamps to its min/max range.
-        """
-        if index is None or len(index) == 0:
-            return None
-        first, last = index.min(), index.max()
-        if isinstance(first, datetime):
-            first = first.date()
-        if isinstance(last, datetime):
-            last = last.date()
-        if not isinstance(first, date) or not isinstance(last, date):
-            return None
-        return self.session_dates(first, last)
-
-    def is_session(self, d: date) -> bool | None:
-        sessions = self.session_dates(d, d)
-        if sessions is None:
-            return None
-        return d in sessions
-
-    # -- schedule -----------------------------------------------------------
-
-    def is_open(self, at: datetime | None = None) -> bool | None:
-        """Whether a *regular* session is running at ``at`` (UTC now default)."""
-        try:
-            return bool(self._cal.is_open_at_time(_as_utc(at)))
-        except Exception:
-            return None
-
-    def next_open(self, at: datetime | None = None) -> datetime | None:
-        return self._schedule_point("next_open", at)
-
-    def next_close(self, at: datetime | None = None) -> datetime | None:
-        return self._schedule_point("next_close", at)
-
-    def previous_close(self, at: datetime | None = None) -> datetime | None:
-        return self._schedule_point("previous_close", at)
-
-    def _schedule_point(self, method: str, at: datetime | None) -> datetime | None:
-        try:
-            return getattr(self._cal, method)(_as_utc(at)).to_pydatetime()
-        except Exception:
-            return None
-
-    def phase(self, at: datetime | None = None) -> str | None:
-        """Trading phase at ``at``: "premarket" | "open" | "aftermarket" | "closed".
-
-        Regular hours come from the calendar; the extended windows are the
-        venue's ``ExtendedHours`` offsets around them. Venues without extended
-        hours only ever report "open"/"closed". This is the *scheduled* phase —
-        the live authority for what a venue is actually doing right now is the
-        quote feed's own market_state; use this for prediction and fallback.
-        """
-        ts = _as_utc(at)
-        try:
-            if self._cal.is_open_at_time(ts):
-                return "open"
-            # previous_close is strictly exclusive: at the exact close instant
-            # it returns the *prior* session's close, which would misfile the
-            # first moment of aftermarket as closed. Nudging the query point
-            # one minute forward makes a close at ts count as "just closed";
-            # mid-session instants can't reach here (is_open returned above).
-            prev_close = self._cal.previous_close(ts + pd.Timedelta(minutes=1))
-            nxt_open = self._cal.next_open(ts)
-        except Exception:
-            return None
-        if self.extended_hours is not None:
-            if ts < prev_close + self.extended_hours.post_offset:
-                return "aftermarket"
-            if ts >= nxt_open - self.extended_hours.pre_offset:
-                return "premarket"
-        return "closed"
-
-
-# One Venue per calendar name, shared by every Symbol that resolves to it.
-_venues: dict[str, Venue | None] = {}
-
-
-def _venue_for(name: str) -> Venue | None:
-    if name not in _venues:
-        try:
-            import exchange_calendars as xcals
-
-            calendar = xcals.get_calendar(name)
-            _venues[name] = Venue(name, calendar, EXTENDED_HOURS.get(name))
-        except Exception:
-            logger.warning("Could not build trading calendar %r", name, exc_info=True)
-            _venues[name] = None
-    return _venues[name]
-
-
-class Symbol(str):
-    """A Yahoo ticker that knows its venue: ``Symbol("IWDA.AS").venue``.
-
-    A ``str`` subclass, so it drops in anywhere a plain symbol string is
-    used (dict keys, comparisons, serialization) — wrap at the point where
-    venue questions arise, no need to thread a new type through the app.
-    Venue resolution is cached on the instance; the Venue itself is shared
-    per calendar. Note that str operations (``.upper()`` etc.) return plain
-    ``str`` — re-wrap if you still need ``.venue`` afterwards.
-    """
-
-    @cached_property
-    def _instrument(self) -> Listing:
-        """One-pass ticker-shape classification — the single place shape is
-        interpreted. Everything else reads the resulting traits."""
-        sym = self.upper().strip()
-        if not sym:
-            return Listing(None, None)
-        if sym in INDEX_CALENDARS:
-            return Listing(INDEX_CALENDARS[sym], None)
-        if sym.startswith("^"):
-            return Listing(None, None)  # unknown index — don't guess a venue
-        if sym.endswith("=X") or sym.endswith("=F"):
-            return Listing(None, None)  # FX / futures — not exchange-session shaped
-        if "-" in sym:
-            quote = sym.rsplit("-", 1)[1]
-            if quote in CRYPTO_QUOTE_CURRENCIES:
-                # Crypto pairs (BTC-USD): every day is a session; the quote
-                # leg is the display currency when it's fiat.
-                return Listing("24/7", quote if quote in _FIAT_QUOTES else None)
-        if "." in sym:
-            return SUFFIX_LISTINGS.get(sym.rsplit(".", 1)[1], Listing(None, None))
-        return Listing(DEFAULT_US_CALENDAR, "USD")
-
-    @property
-    def calendar_name(self) -> str | None:
-        """The exchange_calendars name for this ticker, or None."""
-        return self._instrument.calendar
-
-    @property
-    def currency(self) -> str | None:
-        """The venue/pair currency inferred from ticker shape, or None.
-
-        A fallback only — Yahoo's own currency field wins when present (see
-        ``resolve_currency``); this answers when Yahoo doesn't.
-        """
-        return self._instrument.currency
-
-    @cached_property
-    def venue(self) -> Venue | None:
-        """This ticker's trading venue, or None when it can't be resolved."""
-        name = self.calendar_name
-        return _venue_for(name) if name else None
-
-
-def any_venue_open(symbols, at: datetime | None = None) -> bool:
-    """Whether any symbol's venue is in a tradeable phase (incl. extended hours).
-
-    The scheduler gate: background jobs that only matter while something can
-    trade (intraday bar sync, price heal) ask this instead of weekday guesses
-    or sampling live quotes. Symbols sharing a venue are checked once — a big
-    portfolio collapses to a handful of cached Venue objects and zero API
-    calls.
-
-    Fail-open by design: an unresolvable venue or a failed schedule query
-    counts as tradeable. The gate exists to skip certainly-dead ticks, never
-    to block real work — a wrong True costs one harmless fetch, a wrong False
-    silently stalls data.
-    """
-    seen: set[str] = set()
-    for sym in symbols:
-        s = Symbol(sym)
-        name = s.calendar_name
-        if name is None:
-            return True  # unknown venue — trading can't be ruled out
-        if name in seen:
-            continue
-        seen.add(name)
-        venue = s.venue
-        if venue is None:
-            return True  # calendar failed to build — same fail-open rule
-        phase = venue.phase(at)
-        if phase is None or phase != "closed":
-            return True
-    return False
-
-
-_PHASE_RANK = {"closed": 0, "premarket": 1, "aftermarket": 1, "open": 2}
-
-
-def schedule_poll_hint(symbols, at: datetime | None = None) -> tuple[str, float | None]:
-    """The most-active scheduled phase across the symbols' venues, plus the
-    seconds until the earliest upcoming regular open.
-
-    Feeds the SSE poll cadence: the live quote feed stays authoritative for
-    *speeding up* (it knows about halts and special sessions), while this hint
-    backstops it — a dead quote batch during regular hours must not slow
-    polling — and lets an all-closed stream sleep exactly until the next bell.
-
-    Unlike :func:`any_venue_open`, unresolvable symbols contribute nothing
-    here rather than failing open: a wrong "open" would pin the fast poll
-    forever, whereas contributing nothing merely falls back to the live-state
-    cadence. Returns ``("closed", None)`` when no venue resolves at all.
-    """
-    best = "closed"
-    next_open_secs: float | None = None
-    ts = _as_utc(at).to_pydatetime()
-    seen: set[str] = set()
-    for sym in symbols:
-        s = Symbol(sym)
-        name = s.calendar_name
-        if name is None or name in seen:
-            continue
-        seen.add(name)
-        venue = s.venue
-        if venue is None:
-            continue
-        phase = venue.phase(ts)
-        if phase is not None and _PHASE_RANK.get(phase, 0) > _PHASE_RANK[best]:
-            best = phase
-        nxt = venue.next_open(ts)
-        if nxt is not None:
-            secs = (nxt - ts).total_seconds()
-            if secs > 0 and (next_open_secs is None or secs < next_open_secs):
-                next_open_secs = secs
-    return best, next_open_secs
